+++ src/dom/index.ts
@@ -1,4 +1,4 @@
 export { Attribute } from './attribute';
-export { DOMNode, NodeClosed } from './domnode';
+export { HtmlElement, NodeClosed } from './htmlelement';
 export { DOMTokenList } from './domtokenlist';
 export { DOMTree } from './domtree';