+++ src/rules/button-type.ts
@@ -1,5 +1,5 @@
 import { Rule } from '../rule';
-import { DOMNode } from 'dom';
+import { HtmlElement } from 'dom';
 import { DOMReadyEvent } from '../event';
 
 const validTypes = ['submit', 'button', 'reset'];
@@ -8,7 +8,7 @@ class ButtonType extends Rule {
 	setup(){
 		this.on('dom:ready', (event: DOMReadyEvent) => {
 			const buttons = event.document.getElementsByTagName('button');
-			buttons.forEach((node: DOMNode) => {
+			buttons.forEach((node: HtmlElement) => {
 				const attr = node.getAttribute('type');
 				if (attr === null){
 					this.report(node, "Button is missing type attribute");